split up main.js and dashboard.js
