+++ README.md
@@ -45,4 +45,3 @@ To avoid GitHub API rate limits, configure a personal access token:
 4. Enter your token and save
 
 The dashboard supports both Classic (`ghp_`) and Fine-grained (`github_pat_`) tokens.
-
