+++ index.html
@@ -24,9 +24,7 @@
 
         <!-- Settings Container -->
         <div id="settings-container">
-            <button id="settings-btn" onclick="toggleSettings()">
-                ⚙️  
-            </button>
+            <button id="settings-btn" onclick="toggleSettings()">⚙️</button>
             <div id="settings-panel" style="display: none">
                 <h3>GitHub Authentication</h3>
                 <div class="token-status">
