+++ src/dashboard.js
@@ -13,16 +13,15 @@
 
 import {
     getAuthHeaders,
-    getTriggerInfo,
     isHistoricView,
     getGlDays,
     formatGLDate,
     formatDetailedDate,
-    formatDateTimeDetailed,
     shouldLoadHistoricReleases,
     setElementStatus,
-    bulkSetElementStatus,
     getBranchParameter,
+    calculateTargetDate,
+    calculateHistoricPipelineDuration,
 } from "./utils.js";
 
 import {
@@ -40,6 +39,18 @@ import {
 
 // Import JSZip for artifact extraction
 import JSZip from "jszip";
+import {
+    renderHistoricReleases,
+    updateCurrentReleaseHeaderColors,
+    updateCurrentReleaseSummary,
+    updateStageColor,
+    updatePipelineColor,
+    createRunItemHTML,
+    updateHeaderColor,
+    getRunStatus,
+} from "./ui.js";
+
+import { getParentWorkflowInfo } from "./parentWorkflow.js";
 
 // ========================================
 // GLOBAL STATE MANAGEMENT
@@ -605,32 +616,6 @@ export async function fillPackageTable() {
 // ========================================
 // SHARED WORKFLOW UTILITIES
 // ========================================
-function getRunStatus(run) {
-    let statusClass = "";
-    let statusText = "";
-
-    if (run.status === "in_progress") {
-        statusClass = "progress";
-        statusText = "In Progress";
-    } else if (run.status === "queued") {
-        statusClass = "queued";
-        statusText = "Queued";
-    } else if (run.status === "completed") {
-        if (run.conclusion === "success") {
-            statusClass = "success";
-            statusText = "Success";
-        } else {
-            statusClass = "failure";
-            statusText = run.conclusion || "Failed";
-        }
-    } else {
-        statusClass = "queued";
-        statusText = run.status;
-    }
-
-    return { statusClass, statusText };
-}
-
 function calculateDuration(run) {
     if (run.status !== "completed") return "";
 
@@ -653,115 +638,6 @@ function calculateDuration(run) {
     } else {
         return `${durationSeconds}s`;
     }
-}
-
-async function createRunItemHTML(run, workflow, useFullDate = false) {
-    const { statusClass, statusText } = getRunStatus(run);
-
-    // Check if this is a Stage 4 workflow or platform cleanup
-    const isStage4Workflow = [
-        WORKFLOW_IDS.PUBLISH_GHCR,
-        WORKFLOW_IDS.PUBLISH_S3,
-    ].includes(workflow.id);
-    const isPlatformCleanup =
-        workflow.id === WORKFLOW_IDS.PLATFORM_TEST_CLEANUP;
-
-    // Use detailed date/time format for all stages and platform cleanup
-    const useDetailedDateTime = true; // Always use detailed format now
-
-    const createdTime = useDetailedDateTime
-        ? formatDateTimeDetailed(new Date(run.created_at))
-        : new Date(run.created_at).toLocaleTimeString();
-    const updatedTime = useDetailedDateTime
-        ? formatDateTimeDetailed(new Date(run.updated_at))
-        : new Date(run.updated_at).toLocaleTimeString();
-
-    const durationText = calculateDuration(run);
-    const branch = run.head_branch || "main";
-    const commitSha = run.head_sha ? run.head_sha.substring(0, 7) : "unknown";
-
-    const triggerInfo = await getTriggerInfo(
-        API_CONFIG.GARDENLINUX_ORG,
-        workflow.repo,
-        run
-    );
-
-    let timeDisplay = `Start: ${createdTime}`;
-    let durationDisplay = "";
-
-    if (run.status === "completed") {
-        timeDisplay += ` | End: ${updatedTime}`;
-        if (durationText) {
-            durationDisplay = `Duration: ${durationText}`;
-        }
-    } else if (run.status === "in_progress") {
-        timeDisplay += ` | Running...`;
-    }
-
-    // Check if this is a Stage 4 workflow and attempt to get parent run information
-    let parentRunInfo = null;
-
-    if (isStage4Workflow) {
-        try {
-            console.log(
-                `🔍 [Stage 4] Checking for parent run in workflow ${workflow.name} (${run.id})`
-            );
-            parentRunInfo = await getParentWorkflowInfo(
-                API_CONFIG.GARDENLINUX_ORG,
-                workflow.repo,
-                run.id
-            );
-            if (parentRunInfo && parentRunInfo.parentRunId) {
-                console.log(
-                    `🔍 [Stage 4] Found parent run ID: ${parentRunInfo.parentRunId} for run ${run.id}`
-                );
-            }
-        } catch (error) {
-            console.log(
-                `🔍 [Stage 4] Error getting parent info for run ${run.id}:`,
-                error.message
-            );
-        }
-    }
-
-    // Build the parent run display
-    let parentRunDisplay = "";
-    if (parentRunInfo && parentRunInfo.parentRunId) {
-        parentRunDisplay = `
-            <div class="parent-run-info">
-                parent run: <a href="https://github.com/gardenlinux/gardenlinux/actions/runs/${parentRunInfo.parentRunId}"
-                   target="_blank"
-                   class="parent-run-link"
-                   title="View parent workflow run that triggered this">${parentRunInfo.parentRunId}</a>
-            </div>
-        `;
-    } else if (isStage4Workflow) {
-        // Always show parent run info for Stage 4 workflows, even if not found
-        parentRunDisplay = `
-            <div class="parent-run-info">
-                <span class="parent-run-unavailable" title="${parentRunInfo?.message || "No parent run information available"}">parent run: Not found</span>
-            </div>
-        `;
-    }
-
-    return `
-        <a href="${run.html_url}" target="_blank" class="run-item-link">
-            <div class="run-status-line">
-                <strong class="status-${statusClass}">${statusText}</strong>
-            </div>
-            <div class="run-timing-line">
-                ${timeDisplay}
-            </div>
-            ${durationDisplay ? `<div class="run-duration-line">${durationDisplay}</div>` : ""}
-            <div class="run-meta">
-                <span>Branch: ${branch}</span> |
-                <span>Commit: ${commitSha}</span> |
-                <span>Run: ${run.id}</span> |
-                <span>Trigger: ${triggerInfo}</span>
-            </div>
-            ${parentRunDisplay}
-        </a>
-    `;
 }
 
 // ========================================
@@ -896,261 +772,17 @@ export function updatePipelineHierarchy() {
     updateCurrentReleaseHeaderColors(pipelineStatus);
 
     // Update current release summary
-    updateCurrentReleaseSummary(stageStatuses, pipelineStatus);
+    updateCurrentReleaseSummary(
+        stageStatuses,
+        pipelineStatus,
+        packageStatus,
+        workflowRunData,
+        WORKFLOW_IDS,
+        getGlDays
+    );
 
     console.log("Stage statuses:", stageStatuses);
     console.log("Overall pipeline status:", pipelineStatus);
-}
-
-function updateCurrentReleaseHeaderColors(status) {
-    const releaseHeader = document.getElementById("current-release-header");
-    const detailsHeader = document.getElementById("current-details-header");
-
-    // Update headers with status
-    setElementStatus(releaseHeader, status, "status-");
-    setElementStatus(detailsHeader, status, "status-");
-}
-
-function updateCurrentReleaseSummary(stageStatuses, pipelineStatus) {
-    const glDays = getGlDays();
-    const formattedDate = formatDetailedDate(glDays);
-
-    // Update GL version and date
-    const currentGlVersionElement =
-        document.getElementById("current-gl-version");
-    const currentDateElement = document.getElementById("current-date");
-
-    if (currentGlVersionElement) {
-        currentGlVersionElement.textContent = `GL ${glDays}`;
-        // Add status class to GL version badge
-        setElementStatus(currentGlVersionElement, pipelineStatus);
-    }
-
-    if (currentDateElement) {
-        currentDateElement.textContent = formattedDate;
-    }
-
-    // Update overall status indicator
-    const currentStatusIndicator = document.getElementById(
-        "current-status-indicator"
-    );
-    if (currentStatusIndicator) {
-        setElementStatus(currentStatusIndicator, pipelineStatus);
-        currentStatusIndicator.title = `Overall Status: ${pipelineStatus}`;
-    }
-
-    // Update stage dots
-    for (let i = 1; i <= 4; i++) {
-        const stageDot = document.getElementById(`current-stage-${i}`);
-        if (stageDot) {
-            const stageStatus = stageStatuses[`stage-${i}`] || "unknown";
-            setElementStatus(stageDot, stageStatus);
-        }
-    }
-
-    // Update summary text
-    const currentSummaryElement = document.getElementById("current-summary");
-    if (currentSummaryElement) {
-        if (pipelineStatus === "success") {
-            currentSummaryElement.textContent = "Release successful";
-        } else if (pipelineStatus === "progress") {
-            currentSummaryElement.textContent = "Pipeline in progress";
-        } else if (pipelineStatus === "failure") {
-            currentSummaryElement.textContent = "Pipeline failures";
-        } else if (pipelineStatus === "warning") {
-            currentSummaryElement.textContent = "Issues detected";
-        } else {
-            currentSummaryElement.textContent = "Status loading...";
-        }
-    }
-
-    // Update package status
-    const currentPackageStatusElement = document.getElementById(
-        "current-package-status"
-    );
-    if (currentPackageStatusElement) {
-        if (packageStatus === "warning") {
-            const packageStatusElement =
-                document.getElementById("package-status");
-            const packageText = packageStatusElement
-                ? packageStatusElement.textContent
-                : "Package issues detected";
-            currentPackageStatusElement.textContent = packageText.replace(
-                "packages need attention",
-                "pkg issues"
-            );
-        } else if (packageStatus === "success") {
-            currentPackageStatusElement.textContent = "All packages OK";
-        } else if (packageStatus === "no-data") {
-            currentPackageStatusElement.textContent = "No package data";
-        } else if (packageStatus === "api-error") {
-            currentPackageStatusElement.textContent = "Package data error";
-        } else {
-            currentPackageStatusElement.textContent = "Loading packages...";
-        }
-    }
-
-    // Update duration with calculated pipeline duration
-    const currentDurationElement = document.getElementById("current-duration");
-    if (currentDurationElement) {
-        const duration = calculatePipelineDuration(
-            stageStatuses,
-            pipelineStatus
-        );
-        currentDurationElement.textContent = duration;
-        currentDurationElement.title = "Duration of stages 3 and 4";
-    }
-}
-
-// Helper function to calculate overall pipeline duration
-function calculatePipelineDuration(stageStatuses, pipelineStatus) {
-    // Always try to calculate actual duration from Stage 3 to Stage 4 first
-    const stage3Workflows = [WORKFLOW_IDS.NIGHTLY, WORKFLOW_IDS.MANUAL_RELEASE];
-    const stage4Workflows = [
-        WORKFLOW_IDS.PUBLISH_GHCR,
-        WORKFLOW_IDS.PUBLISH_S3,
-    ];
-    let earliestStage3Start = null;
-    let latestStage4End = null;
-    let hasInProgressWorkflows = false;
-
-    // Find earliest Stage 3 start time and check for in-progress workflows
-    for (const workflowId of stage3Workflows) {
-        if (workflowRunData && workflowRunData[workflowId]) {
-            const runData = workflowRunData[workflowId];
-
-            // Check if workflow is in progress
-            if (
-                runData.status === "in_progress" ||
-                runData.status === "queued"
-            ) {
-                hasInProgressWorkflows = true;
-            }
-
-            // Get start time from completed or in-progress workflows
-            if (
-                runData.status === "completed" ||
-                runData.status === "in_progress" ||
-                runData.status === "queued"
-            ) {
-                const startTime = new Date(runData.created_at);
-                if (!earliestStage3Start || startTime < earliestStage3Start) {
-                    earliestStage3Start = startTime;
-                }
-            }
-        }
-    }
-
-    // Find latest Stage 4 end time and check for in-progress workflows
-    for (const workflowId of stage4Workflows) {
-        if (workflowRunData && workflowRunData[workflowId]) {
-            const runData = workflowRunData[workflowId];
-
-            // Check if workflow is in progress
-            if (
-                runData.status === "in_progress" ||
-                runData.status === "queued"
-            ) {
-                hasInProgressWorkflows = true;
-            }
-
-            // Only use end time from completed workflows
-            if (runData.status === "completed") {
-                const endTime = new Date(runData.updated_at);
-                if (!latestStage4End || endTime > latestStage4End) {
-                    latestStage4End = endTime;
-                }
-            }
-        }
-    }
-
-    // If we have a start time, calculate duration
-    if (earliestStage3Start) {
-        let endTime;
-        let isRunning = false;
-
-        // If there are in-progress workflows or no completed Stage 4, use current time
-        if (hasInProgressWorkflows || !latestStage4End) {
-            endTime = new Date(); // Current time
-            isRunning = true;
-        } else {
-            endTime = latestStage4End; // Completed pipeline
-        }
-
-        const durationMs = endTime - earliestStage3Start;
-        if (durationMs > 0) {
-            const durationHours = Math.floor(durationMs / 3600000);
-            const durationMinutes = Math.floor((durationMs % 3600000) / 60000);
-
-            let durationText;
-            if (durationHours > 0) {
-                durationText = `${durationHours}h ${durationMinutes}m`;
-            } else {
-                durationText = `${durationMinutes}m`;
-            }
-
-            // Add "(running)" indicator if pipeline is still in progress
-            if (isRunning) {
-                durationText += " (running)";
-            }
-
-            return durationText;
-        }
-    }
-
-    // Fallback to status-based messages only if duration calculation isn't possible
-    if (pipelineStatus === "progress" || pipelineStatus === "unknown") {
-        return "In progress...";
-    } else if (pipelineStatus === "failure") {
-        return "Pipeline failed";
-    } else if (pipelineStatus === "warning") {
-        return "Issues detected";
-    }
-
-    return "Completed";
-}
-
-// ========================================
-// COLOR MANAGEMENT & VISUAL UPDATES
-// ========================================
-function updateStageColor(stageId, status) {
-    const stage = document.getElementById(stageId);
-    setElementStatus(stage, status, "stage-");
-}
-
-function updatePipelineColor(status) {
-    const pipelineContainer = document.getElementById("pipeline-container");
-    setElementStatus(pipelineContainer, status, "pipeline-");
-}
-
-export function updateHeaderColor(status) {
-    const headerElement = document.getElementById("gl-days");
-    if (!headerElement) return;
-
-    // Remove existing header status classes
-    headerElement.classList.remove(
-        "header-success",
-        "header-failure",
-        "header-progress",
-        "header-warning",
-        "header-unknown"
-    );
-
-    // Map statuses to header classes
-    const statusMap = {
-        success: "header-success",
-        failure: "header-failure",
-        progress: "header-progress",
-        warning: "header-warning",
-        unknown: "header-unknown",
-        error: "header-failure", // Map error to failure for header
-        "no-data": "header-unknown",
-    };
-
-    const headerClass = statusMap[status];
-    if (headerClass) {
-        headerElement.classList.add(headerClass);
-    }
 }
 
 // ========================================
@@ -1224,73 +856,6 @@ export async function loadHistoricReleases() {
     }
 }
 
-function calculateTargetDate(glDays) {
-    const initialDay = new Date(GL_INITIAL_DATE);
-    const targetDate = new Date(initialDay);
-    targetDate.setDate(targetDate.getDate() + glDays);
-    targetDate.setHours(0, 0, 0, 0);
-    return targetDate;
-}
-
-// Helper function to calculate historic pipeline duration
-function calculateHistoricPipelineDuration(workflowRunData) {
-    const stage3Workflows = [WORKFLOW_IDS.NIGHTLY, WORKFLOW_IDS.MANUAL_RELEASE];
-    const stage4Workflows = [
-        WORKFLOW_IDS.PUBLISH_GHCR,
-        WORKFLOW_IDS.PUBLISH_S3,
-    ];
-    let earliestStage3Start = null;
-    let latestStage4End = null;
-
-    // Find earliest Stage 3 start time
-    for (const workflowId of stage3Workflows) {
-        if (workflowRunData && workflowRunData[workflowId]) {
-            const runData = workflowRunData[workflowId];
-            if (
-                runData.status === "completed" ||
-                runData.status === "in_progress" ||
-                runData.status === "queued"
-            ) {
-                const startTime = new Date(runData.created_at);
-                if (!earliestStage3Start || startTime < earliestStage3Start) {
-                    earliestStage3Start = startTime;
-                }
-            }
-        }
-    }
-
-    // Find latest Stage 4 end time (only from completed workflows for historic data)
-    for (const workflowId of stage4Workflows) {
-        if (workflowRunData && workflowRunData[workflowId]) {
-            const runData = workflowRunData[workflowId];
-            if (runData.status === "completed") {
-                const endTime = new Date(runData.updated_at);
-                if (!latestStage4End || endTime > latestStage4End) {
-                    latestStage4End = endTime;
-                }
-            }
-        }
-    }
-
-    // Calculate duration if we have both start and end times
-    if (earliestStage3Start && latestStage4End) {
-        const durationMs = latestStage4End - earliestStage3Start;
-        if (durationMs > 0) {
-            const durationHours = Math.floor(durationMs / 3600000);
-            const durationMinutes = Math.floor((durationMs % 3600000) / 60000);
-
-            if (durationHours > 0) {
-                return `${durationHours}h ${durationMinutes}m`;
-            } else {
-                return `${durationMinutes}m`;
-            }
-        }
-    }
-
-    // Return null if duration can't be calculated
-    return null;
-}
-
 async function loadHistoricDay(glDays) {
     try {
         // Get basic info
@@ -1305,7 +870,10 @@ async function loadHistoricDay(glDays) {
         const workflowRunData = workflowResult.workflowRunData;
 
         // Calculate pipeline duration
-        const duration = calculateHistoricPipelineDuration(workflowRunData);
+        const duration = calculateHistoricPipelineDuration(
+            workflowRunData,
+            WORKFLOW_IDS
+        );
 
         return {
             glDays,
@@ -1370,7 +938,7 @@ async function getHistoricWorkflowStatus(glDays) {
     // Store actual run data for duration calculation
     const workflowRunData = {};
 
-    const targetDate = calculateTargetDate(glDays);
+    const targetDate = calculateTargetDate(glDays, GL_INITIAL_DATE);
     const nextDay = new Date(targetDate);
     nextDay.setDate(nextDay.getDate() + 1);
 
@@ -1664,359 +1232,4 @@ function calculateOverallStatus(packageStatus, workflowStatus) {
 
     // Default to unknown for any other combination
     return "unknown";
-}
-
-function renderHistoricReleases(historicData) {
-    const historicList = document.getElementById("historic-releases-list");
-
-    if (historicData.length === 0) {
-        historicList.innerHTML = `
-            <div class="historic-loading">
-                No historic data available for the last 14 days.
-            </div>
-        `;
-        return;
-    }
-
-    historicList.innerHTML = historicData
-        .map(
-            (day) => `
-        <a href="?gl=${day.glDays}&no_historic_releases=true" target="_blank" class="historic-release-row" title="View detailed dashboard for GL ${day.glDays}">
-            <div class="historic-gl-version ${day.overallStatus}">GL ${day.glDays}</div>
-            <div class="historic-date">${day.date}</div>
-
-            <div class="historic-overall-status">
-                <span class="historic-status-indicator ${day.overallStatus}"
-                      title="Overall Status: ${day.overallStatus}"></span>
-            </div>
-
-            <div class="historic-stages" title="Stages: Package | Repo | Build | Publish">
-                <span class="historic-stage-dot ${getStageColorClass(day.packageStatus)}" title="Package Builds"></span>
-                <span class="historic-stage-dot ${day.workflowStatus?.["stage-2"] || "unknown"}" title="Repository"></span>
-                <span class="historic-stage-dot ${day.workflowStatus?.["stage-3"] || "unknown"}" title="Build & Release"></span>
-                <span class="historic-stage-dot ${day.workflowStatus?.["stage-4"] || "unknown"}" title="Publish"></span>
-            </div>
-
-            <div class="historic-package-status">
-                ${
-                    day.packageStatus.status === "success"
-                        ? "All packages OK"
-                        : day.packageStatus.status === "warning"
-                          ? `${day.packageStatus.issueCount || 0} pkg issues`
-                          : day.packageStatus.status === "error"
-                            ? "Package data error"
-                            : day.packageStatus.status === "loading"
-                              ? "Loading packages..."
-                              : "No package data"
-                }
-            </div>
-
-            <div class="historic-duration" title="Duration of stages 3 and 4">
-                ${day.duration || "No data"}
-            </div>
-
-            <div class="historic-summary">
-                ${
-                    day.overallStatus === "success"
-                        ? "Release successful"
-                        : day.overallStatus === "progress"
-                          ? "Pipeline in progress"
-                          : day.overallStatus === "failure"
-                            ? "Pipeline failures"
-                            : day.overallStatus === "warning"
-                              ? "Issues detected"
-                              : "Status loading..."
-                }
-            </div>
-        </a>
-    `
-        )
-        .join("");
-}
-
-// Helper function to convert package status to stage color class
-function getStageColorClass(packageStatus) {
-    switch (packageStatus.status) {
-        case "success":
-            return "success";
-        case "warning":
-            return "warning";
-        case "error":
-            return "failure";
-        case "no-data":
-            return "unknown";
-        default:
-            return "unknown";
-    }
-}
-
-// ========================================
-// PARENT WORKFLOW UTILITIES
-// ========================================
-
-/**
- * Download and extract artifact data to find parent workflow information
- */
-async function downloadAndExtractArtifact(owner, repo, artifact) {
-    try {
-        console.log(
-            `🔍 [DEBUG] Attempting to download artifact ${artifact.name} (ID: ${artifact.id})...`
-        );
-
-        // Check if artifact name is in allowed list
-        const isAllowed = ALLOWED_ARTIFACT_NAMES.some((allowedName) =>
-            artifact.name.toLowerCase().includes(allowedName.toLowerCase())
-        );
-
-        if (!isAllowed) {
-            console.log(
-                `🔍 [DEBUG] Artifact ${artifact.name} not in allowed list, skipping download`
-            );
-            return {
-                success: false,
-                reason: "not_allowed",
-                message: `Artifact name '${artifact.name}' not in allowed list: ${ALLOWED_ARTIFACT_NAMES.join(", ")}`,
-            };
-        }
-
-        const downloadResponse = await fetch(
-            `${API_CONFIG.GITHUB_API_BASE}/repos/${owner}/${repo}/actions/artifacts/${artifact.id}/zip`,
-            {
-                headers: getAuthHeaders(),
-            }
-        );
-
-        if (!downloadResponse.ok) {
-            const status = downloadResponse.status;
-            if (status === 401 || status === 403) {
-                return {
-                    success: false,
-                    reason: "auth_required",
-                    message: `${status}: Authentication required`,
-                    status: status,
-                };
-            } else {
-                return {
-                    success: false,
-                    reason: "download_failed",
-                    message: `${status}: ${downloadResponse.statusText}`,
-                    status: status,
-                };
-            }
-        }
-
-        console.log(
-            `🔍 [DEBUG] Successfully downloaded artifact ${artifact.name}`
-        );
-
-        const arrayBuffer = await downloadResponse.arrayBuffer();
-        const zip = new JSZip();
-        const loadedZip = await zip.loadAsync(arrayBuffer);
-
-        console.log(
-            `🔍 [DEBUG] ZIP file loaded, contains ${Object.keys(loadedZip.files).length} files`
-        );
-
-        // Extract and parse files
-        const extractedData = {};
-        let jobId = null;
-        let parentRunId = null;
-
-        for (const [filename, file] of Object.entries(loadedZip.files)) {
-            if (!file.dir) {
-                console.log(`🔍 [DEBUG] Extracting file: ${filename}`);
-                const content = await file.async("text");
-                extractedData[filename] = content;
-
-                // Try to parse as JSON if possible
-                try {
-                    if (filename.toLowerCase().endsWith(".json")) {
-                        const jsonData = JSON.parse(content);
-                        extractedData[filename + "_parsed"] = jsonData;
-
-                        // Look for id in the JSON data
-                        if (jsonData.id && !parentRunId) {
-                            parentRunId = jsonData.id;
-                            console.log(
-                                `🔍 [DEBUG] Found parent run id in ${filename}: ${parentRunId}`
-                            );
-                        }
-                    }
-                } catch (parseError) {
-                    console.log(
-                        `🔍 [DEBUG] Could not parse ${filename} as JSON`
-                    );
-                }
-            }
-        }
-
-        return {
-            success: true,
-            extractedData: extractedData,
-            jobId: jobId,
-            parentRunId: parentRunId,
-            fileCount: Object.keys(extractedData).length,
-            message: `Successfully extracted ${Object.keys(extractedData).length} files`,
-        };
-    } catch (error) {
-        console.error(
-            `🔍 [DEBUG] Error downloading/extracting artifact:`,
-            error
-        );
-        return {
-            success: false,
-            reason: "extraction_error",
-            message: error.message,
-            error: error,
-        };
-    }
-}
-
-/**
- * Get parent workflow information and artifacts
- */
-async function getParentWorkflowInfo(owner, repo, runId) {
-    try {
-        console.log(`🔍 [DEBUG] Fetching artifacts for run ${runId}...`);
-
-        const artifactsResponse = await fetch(
-            `${API_CONFIG.GITHUB_API_BASE}/repos/${owner}/${repo}/actions/runs/${runId}/artifacts`,
-            {
-                headers: getAuthHeaders(),
-            }
-        );
-
-        if (!artifactsResponse.ok) {
-            return {
-                found: false,
-                message: `Failed to fetch artifacts: ${artifactsResponse.status} ${artifactsResponse.statusText}`,
-                error: `API Error: ${artifactsResponse.status}`,
-            };
-        }
-
-        const artifactsData = await artifactsResponse.json();
-        const artifacts = artifactsData.artifacts || [];
-
-        console.log(
-            `🔍 [DEBUG] Found ${artifacts.length} artifacts for run ${runId}`
-        );
-
-        if (artifacts.length === 0) {
-            return {
-                found: false,
-                message: "No artifacts found for this run",
-            };
-        }
-
-        // Enhanced parent workflow artifact detection
-        // Look for artifacts that likely contain parent workflow information
-        const parentWorkflowArtifacts = artifacts.filter(
-            (artifact) =>
-                artifact.name &&
-                // Check if artifact name is in allowed list for download
-                ALLOWED_ARTIFACT_NAMES.some((allowedName) =>
-                    artifact.name
-                        .toLowerCase()
-                        .includes(allowedName.toLowerCase())
-                )
-        );
-
-        console.log(
-            `🔍 [DEBUG] Found ${parentWorkflowArtifacts.length} allowed parent workflow artifacts:`,
-            parentWorkflowArtifacts.map((a) => a.name)
-        );
-
-        // Try to download and extract allowed artifacts first
-        if (parentWorkflowArtifacts.length > 0) {
-            for (const artifact of parentWorkflowArtifacts) {
-                const extractionResult = await downloadAndExtractArtifact(
-                    owner,
-                    repo,
-                    artifact
-                );
-
-                if (extractionResult.success && extractionResult.parentRunId) {
-                    return {
-                        found: true,
-                        message:
-                            "Parent run ID extracted from downloaded artifact",
-                        artifactId: artifact.id,
-                        artifactName: artifact.name,
-                        parentRunId: extractionResult.parentRunId.toString(),
-                        jobId: extractionResult.jobId,
-                        extractionMethod: "artifact_download_extraction",
-                        extractedData: extractionResult.extractedData,
-                        fileCount: extractionResult.fileCount,
-                    };
-                } else if (extractionResult.success) {
-                    // Downloaded successfully but no parent run ID found
-                    console.log(
-                        `🔍 [DEBUG] Downloaded ${artifact.name} but no parent run ID found in content`
-                    );
-                } else {
-                    // Download failed
-                    console.log(
-                        `🔍 [DEBUG] Failed to download ${artifact.name}: ${extractionResult.message}`
-                    );
-                }
-            }
-        }
-
-        // Enhanced detection for all artifacts that might contain parent workflow information
-        const allParentWorkflowArtifacts = artifacts.filter(
-            (artifact) =>
-                artifact.name && artifact.name === "parent-workflow-data"
-        );
-
-        if (allParentWorkflowArtifacts.length > 0) {
-            for (const artifact of allParentWorkflowArtifacts) {
-                const extractionResult = await downloadAndExtractArtifact(
-                    owner,
-                    repo,
-                    artifact
-                );
-
-                if (extractionResult.success && extractionResult.parentRunId) {
-                    return {
-                        found: true,
-                        message:
-                            "Parent run ID extracted from downloaded artifact",
-                        artifactId: artifact.id,
-                        artifactName: artifact.name,
-                        parentRunId: extractionResult.parentRunId.toString(),
-                        jobId: extractionResult.jobId,
-                        extractionMethod: "artifact_download_extraction",
-                        extractedData: extractionResult.extractedData,
-                        fileCount: extractionResult.fileCount,
-                    };
-                } else if (extractionResult.success) {
-                    // Downloaded successfully but no parent run ID found
-                    console.log(
-                        `🔍 [DEBUG] Downloaded ${artifact.name} but no parent run ID found in content`
-                    );
-                } else {
-                    // Download failed
-                    console.log(
-                        `🔍 [DEBUG] Failed to download ${artifact.name}: ${extractionResult.message}`
-                    );
-                }
-            }
-
-            return {
-                found: false,
-                message: `Found no allowed artifacts`,
-                artifactCount: artifacts.length,
-                availableArtifacts: artifacts.slice(0, 10).map((a) => a.name), // Show first 10 artifact names
-                extractionMethod: "no_parent_indicators",
-            };
-        }
-    } catch (error) {
-        console.error(`🔍 [DEBUG] Error fetching parent workflow info:`, error);
-        return {
-            found: false,
-            message: "Error fetching artifact information",
-            error: error.message,
-            extractionMethod: "error",
-        };
-    }
 }
